pulled out shared table
