+++ jupyterlab_bigquery/src/components/details_panel/table_preview.tsx
@@ -1,44 +1,19 @@
 import * as React from 'react';
 import { stylesheet } from 'typestyle';
 
-import {
-  Table,
-  TableBody,
-  TableHead,
-  TableCell,
-  TableRow,
-  TablePagination,
-} from '@material-ui/core';
-
 import LoadingPanel from '../loading_panel';
 import {
   TableDetailsService,
   TablePreview,
 } from './service/list_table_details';
+import { BQTable } from '../shared/bq_table';
 
 const localStyles = stylesheet({
-  tableHeader: {
-    backgroundColor: '#f0f0f0',
-    color: 'black',
-  },
-  tableCell: {
-    border: 'var(--jp-border-width) solid var(--jp-border-color2)',
-    whiteSpace: 'nowrap',
-  },
-  pagination: {
-    backgroundColor: 'white',
-    fontSize: '13px',
-  },
   previewBody: {
     flex: 1,
     minHeight: 0,
     display: 'flex',
     flexDirection: 'column',
-  },
-  scrollable: {
-    flex: 1,
-    minHeight: 0,
-    overflow: 'auto',
   },
 });
 
@@ -51,8 +26,6 @@ interface Props {
 interface State {
   hasLoaded: boolean;
   isLoading: boolean;
-  page: number;
-  rowsPerPage: number;
   preview: TablePreview;
 }
 
@@ -64,8 +37,6 @@ export default class TablePreviewPanel extends React.Component<Props, State> {
     this.state = {
       hasLoaded: false,
       isLoading: false,
-      page: 0,
-      rowsPerPage: 100,
       preview: { fields: [], rows: [] },
     };
   }
@@ -103,21 +74,8 @@ export default class TablePreviewPanel extends React.Component<Props, State> {
     }
   }
 
-  handleChangePage(event, newPage) {
-    this.setState({ page: newPage });
-  }
-
-  handleChangeRowsPerPage(event) {
-    this.setState({
-      rowsPerPage: parseInt(event.target.value, 10),
-    });
-    this.setState({ page: 0 });
-  }
-
   render() {
-    const { rowsPerPage, page } = this.state;
-    const rows = this.state.preview.rows;
-    const fields = ['Row', ...this.state.preview.fields];
+    const { rows, fields } = this.state.preview;
     if (this.state.isLoading) {
       return <LoadingPanel />;
     } else {
@@ -125,58 +83,11 @@ export default class TablePreviewPanel extends React.Component<Props, State> {
         <div className={localStyles.previewBody}>
           {rows.length > 0 ? (
             <div className={localStyles.previewBody}>
-              <div className={localStyles.scrollable}>
-                <Table
-                  size="small"
-                  style={{ width: 'auto', tableLayout: 'auto' }}
-                >
-                  <TableHead className={localStyles.tableHeader}>
-                    <TableRow>
-                      {fields.map((field, index) => (
-                        <TableCell key={'field_' + index}>{field}</TableCell>
-                      ))}
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {rows
-                      .slice(
-                        page * rowsPerPage,
-                        page * rowsPerPage + rowsPerPage
-                      )
-                      .map((row, indexRow) => (
-                        <TableRow key={'preview_row_' + indexRow}>
-                          <TableCell>
-                            {page * rowsPerPage + indexRow + 1}
-                          </TableCell>
-                          {row.map((cell, indexCell) => (
-                            <TableCell
-                              className={localStyles.tableCell}
-                              key={
-                                'preview_row_' + indexRow + '_cell' + indexCell
-                              }
-                            >
-                              {cell}
-                            </TableCell>
-                          ))}
-                        </TableRow>
-                      ))}
-                  </TableBody>
-                </Table>
-              </div>
-              <TablePagination
-                className={localStyles.pagination}
-                rowsPerPageOptions={[10, 30, 50, 100, 200]}
-                component="div"
-                count={rows.length}
-                rowsPerPage={rowsPerPage}
-                page={page}
-                onChangePage={this.handleChangePage.bind(this)}
-                onChangeRowsPerPage={this.handleChangeRowsPerPage.bind(this)}
-              />
+              <BQTable rows={rows} fields={fields} />
             </div>
           ) : (
-            <div>This table is empty.</div>
-          )}
+              <div>This table is empty.</div>
+            )}
         </div>
       );
     }
