+++ jupyterlab_bigquery/src/components/query_editor/query_editor_tab/query_editor_results.tsx
@@ -1,53 +1,19 @@
 // @ts-nocheck
 import React, { Component } from 'react';
 import { stylesheet } from 'typestyle';
-import {
-  Table,
-  TableHead,
-  TableRow,
-  TableCell,
-  TableBody,
-  TablePagination,
-  IconButton,
-} from '@material-ui/core';
-import {
-  KeyboardArrowLeft,
-  KeyboardArrowRight,
-  FirstPage,
-  LastPage,
-} from '@material-ui/icons';
 import { connect } from 'react-redux';
 import { QueryResult } from '../query_text_editor/query_text_editor';
 import { QueryId } from '../../../reducers/queryEditorTabSlice';
+import { Header } from '../../shared/header';
+import { BQTable } from '../../shared/bq_table';
 
 const localStyles = stylesheet({
-  header: {
-    borderBottom: 'var(--jp-border-width) solid var(--jp-border-color2)',
-    fontSize: '18px',
-    margin: 0,
-    padding: '8px 12px 8px 24px',
-  },
-  tableHeader: {
-    backgroundColor: '#f0f0f0',
-    color: 'black',
-  },
-  tableCell: {
-    border: 'var(--jp-border-width) solid var(--jp-border-color2)',
-  },
-  pagination: {
-    position: 'fixed',
-    bottom: 30,
-    backgroundColor: 'white',
-    fontSize: '13px',
-  },
-  paginationOptions: {
-    display: 'flex',
-    fontSize: '13px',
-  },
   resultsContainer: {
-    overflow: 'auto',
-    width: '100%',
-    height: '-webkit-calc(50% - 50px)',
+    // 5/9 of panel height (in relation to editor)
+    flex: 5,
+    minHeight: 0,
+    display: 'flex',
+    flexDirection: 'column',
   },
 });
 
@@ -59,67 +25,6 @@ interface QueryResultsState {
 interface QueryResultsProps {
   queryResult: QueryResult;
   queryId: QueryId;
-}
-
-interface TablePaginationActionsProps {
-  count: number;
-  page: number;
-  rowsPerPage: number;
-  onChangePage: (
-    event: React.MouseEvent<HTMLButtonElement>,
-    newPage: number
-  ) => void;
-}
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-  const { count, page, rowsPerPage, onChangePage } = props;
-
-  const handleFirstPageButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, 0);
-  };
-
-  const handleBackButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, page - 1);
-  };
-
-  const handleNextButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (
-    event: React.MouseEvent<HTMLButtonElement>
-  ) => {
-    onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <div className={localStyles.paginationOptions}>
-      <IconButton onClick={handleFirstPageButtonClick} disabled={page === 0}>
-        <FirstPage />
-      </IconButton>
-      <IconButton onClick={handleBackButtonClick} disabled={page === 0}>
-        <KeyboardArrowLeft />
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-      >
-        <KeyboardArrowRight />
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-      >
-        <LastPage />
-      </IconButton>
-    </div>
-  );
 }
 
 class QueryResults extends Component<QueryResultsProps, QueryResultsState> {
@@ -134,63 +39,14 @@ class QueryResults extends Component<QueryResultsProps, QueryResultsState> {
     this.queryId = props.queryId;
   }
 
-  handleChangePage(event, newPage) {
-    this.setState({ page: newPage });
-  }
-
-  handleChangeRowsPerPage(event) {
-    this.setState({ rowsPerPage: parseInt(event.target.value, 10) });
-    this.setState({ page: 0 });
-  }
-
   render() {
-    const fields = ['Row', ...this.props.queryResult.labels];
+    const fields = this.props.queryResult.labels;
     const rows = this.props.queryResult.content;
-
-    const { rowsPerPage, page } = this.state;
 
     return (
       <div className={localStyles.resultsContainer}>
-        <div className={localStyles.header}>Query results</div>
-        <Table size="small" style={{ width: 'auto', tableLayout: 'auto' }}>
-          <TableHead className={localStyles.tableHeader}>
-            <TableRow>
-              {this.props.queryResult.labels.length !== 0 &&
-                fields.map((field, index) => (
-                  <TableCell key={`result_field_${index}`}>{field}</TableCell>
-                ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {rows
-              .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-              .map((row, rowIndex) => (
-                <TableRow key={'result_row' + rowIndex}>
-                  <TableCell>{page * rowsPerPage + rowIndex + 1}</TableCell>
-                  {row.map((cell, cellIndex) => (
-                    <TableCell
-                      className={localStyles.tableCell}
-                      key={'result_row_' + rowIndex + '_cell' + cellIndex}
-                    >
-                      {cell}
-                    </TableCell>
-                  ))}
-                </TableRow>
-              ))}
-          </TableBody>
-        </Table>
-        {/* TODO(cxjia): hide table pagination when result rows <= 10 */}
-        <TablePagination
-          className={localStyles.pagination}
-          rowsPerPageOptions={[10, 30, 50, 100, 200]}
-          component="div"
-          count={rows.length}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          onChangePage={this.handleChangePage.bind(this)}
-          onChangeRowsPerPage={this.handleChangeRowsPerPage.bind(this)}
-          ActionsComponent={TablePaginationActions}
-        />
+        <Header text="Query results" />
+        {fields.length > 0 ? <BQTable fields={fields} rows={rows} /> : <></>}
       </div>
     );
   }
