+++ jupyterlab_bigquery/src/components/query_editor/query_editor_tab/query_editor_tab.tsx
@@ -24,7 +24,7 @@ class QueryEditorTab extends React.Component<QueryEditorTabProps, {}> {
 
   render() {
     return (
-      <div style={{ height: '100%', width: '100%' }}>
+      <div style={{ display: 'flex', flexDirection: 'column', overflow: 'hidden', height: '100%' }}>
         <QueryTextEditor queryId={this.queryId} />
         <QueryResults queryId={this.queryId} />
       </div>
